+++ MegaManLofi/IReadOnlyArena.h
@@ -18,18 +18,18 @@ namespace MegaManLofi
       virtual const std::shared_ptr<IReadOnlyEntity> GetEntity( int index ) const = 0;
       virtual int GetEntityCount() const = 0;
 
-      virtual long long GetWidth() const = 0;
-      virtual long long GetHeight() const = 0;
+      virtual float GetWidth() const = 0;
+      virtual float GetHeight() const = 0;
 
-      virtual long long GetTileWidth() const = 0;
-      virtual long long GetTileHeight() const = 0;
+      virtual float GetTileWidth() const = 0;
+      virtual float GetTileHeight() const = 0;
 
       virtual int GetHorizontalTiles() const = 0;
       virtual int GetVerticalTiles() const = 0;
 
-      virtual const Quad<long long>& GetActiveRegion() const = 0;
+      virtual const Quad<float>& GetActiveRegion() const = 0;
 
-      virtual const ArenaTile& GetTile( long long index ) const = 0;
+      virtual const ArenaTile& GetTile( int index ) const = 0;
 
       virtual bool HasEntity( int uniqueId ) const = 0;
    };
