+++ MegaManLofi/PlayerThwipInConsoleAnimation.cpp
@@ -89,7 +89,7 @@ void PlayerThwipInConsoleAnimation::Tick()
    else
    {
       _renderDefs->SpriteDefs->PlayerThwipSprite->Tick();
-      auto topDelta = (long long)( _renderDefs->PlayerThwipVelocity * _frameRateProvider->GetFrameSeconds() );
+      auto topDelta = _renderDefs->PlayerThwipVelocity * _frameRateProvider->GetFrameSeconds();
 
       if ( _endPositionChars.Top < _startPositionChars.Top )
       {
